requirements-dev.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.7
Django==4.2.1
django-auditlog==2.2.2
django-cors-headers==4.0.0
django-filter==23.2
django-formtools==2.4
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-rest-passwordreset==1.3.0
django-two-factor-auth==1.15.2
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-standardized-errors==0.12.5
frozenlist==1.3.3
idna==3.4
install==1.3.5
multidict==6.0.4
mysqlclient==2.1.1
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
qrcode==7.4.2
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
black==23.3.0
twilio==8.2.0
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.16.1
urllib3==2.0.2
user-agents==2.2.0
yarl==1.9.2
flake8==6.0.0
black==23.3.0
